image-inpaint/requirements.txt
numpy
torch
torchvision
kornia==0.5.0
opencv-python
Pillow
tqdm
gradio==3.41.2
transformers
diffusers
matplotlib
requests